+++ package.json
@@ -2,29 +2,6 @@
   "name": "amt-tiltaksarrangor-flate",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@babel/core": "^7.26.0",
-    "@navikt/aksel-icons": "^7.9.1",
-    "@navikt/ds-css": "^7.9.1",
-    "@navikt/ds-react": "^7.9.1",
-    "@navikt/ds-tokens": "^7.9.1",
-    "@navikt/nav-dekoratoren-moduler": "^2.1.6",
-    "@sentry/react": "^8.48.0",
-    "@sentry/tracing": "^7.120.3",
-    "@types/react": "^18.3.13",
-    "@types/react-collapse": "^5.0.4",
-    "@types/react-dom": "^18.3.1",
-    "@types/react-router-dom": "^5.3.3",
-    "axios": "^1.7.9",
-    "classnames": "^2.5.1",
-    "constate": "^3.3.2",
-    "dayjs": "^1.11.13",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^7.1.1",
-    "typescript": "^5.7.3",
-    "zod": "^3.24.1"
-  },
   "scripts": {
     "start": "vite",
     "build": "vite build",
@@ -48,6 +25,29 @@
       "prettier --write src/**/*"
     ]
   },
+  "dependencies": {
+    "@babel/core": "^7.26.0",
+    "@navikt/aksel-icons": "^7.9.1",
+    "@navikt/ds-css": "^7.9.1",
+    "@navikt/ds-react": "^7.9.1",
+    "@navikt/ds-tokens": "^7.9.1",
+    "@navikt/nav-dekoratoren-moduler": "^2.1.6",
+    "@sentry/react": "^8.48.0",
+    "@sentry/tracing": "^7.120.3",
+    "@types/react": "^18.3.13",
+    "@types/react-collapse": "^5.0.4",
+    "@types/react-dom": "^18.3.1",
+    "@types/react-router-dom": "^5.3.3",
+    "axios": "^1.7.9",
+    "classnames": "^2.5.1",
+    "constate": "^3.3.2",
+    "dayjs": "^1.11.13",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.1",
+    "typescript": "^5.7.3",
+    "zod": "^3.24.1"
+  },
   "devDependencies": {
     "@faker-js/faker": "^9.3.0",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
@@ -68,7 +68,7 @@
     "eslint-plugin-react-refresh": "^0.4.16",
     "husky": "^9.1.7",
     "lint-staged": "^15.3.0",
-    "msw": "^1.3.5",
+    "msw": "^2.7.0",
     "prettier": "^3.4.2",
     "sass": "^1.83.1",
     "table": "^6.9.0",
@@ -77,6 +77,8 @@
     "vitest": "^2.1.8"
   },
   "msw": {
-    "workerDirectory": "public"
+    "workerDirectory": [
+      "public"
+    ]
   }
 }
